got the form working, and added a few things
I think with the way I set up the header backgrounds, they technically take up the entire viewport. Because of that, inside the form CSS, we had to set position:relative, which I believe brings the element to the front of the page.

Also, I went ahead and set up the checkout page to display the item and its info next to the form. 

Finally, I added a drop down menu that lets you choose what product you're buying, as I just noticed that as a requirement for this project. When you choose an item, the javascript file loads the correct info without reloading the entire html document.